+++ src/core/error/index.ts
@@ -9,6 +9,7 @@ export const errorkeys = {
     'TwoParamsChosen': 'TWO_PARAMS_CHOSEN',
 
     'NodeVersionError': 'NODE_VERSION_ERR',
+    'NodejsMissingPackage': 'NODEJS_MISSING_PKG',
 
     'LengthMaxReached': 'LENGTH_MAX_REACHED',
 
@@ -26,6 +27,7 @@ const messages = {
     [errorkeys.TwoParamsChosen]: () => 'Two parameters has been set to an undefined value, expect one to be undefined value.',
 
     [errorkeys.NodeVersionError]: (versionRequired: string) => 'Your Node.js version is currently under the required version ' + versionRequired + '. Please install a version higher than the installed one.',
+    [errorkeys.NodejsMissingPackage]: (pkg: string) => 'Couldn\'t find the module/package \'' + pkg + '\'. Use \'npm list ' + pkg + '\' to check if it\'s installed or not.',
 
     [errorkeys.LengthMaxReached]: (variableandlength: string) => `The variable/property ${variableandlength[0]} has reached the max length of ${variableandlength[1]}.`,