+++ src/func/idGen.ts
@@ -5,6 +5,6 @@
  * @returns {number}
  */
 
-export function idGen() {
+export function idGen(): number {
     return (Date.now() + Math.floor(Math.random() * 999999999999));
 };